Create a test suite file per package
per @onsi, a package with nested subpackages needs to have a test suite
file per package. Fairly straightforward change to main.go, but in the
process some unnecessary code was cleaned up. A small change to the test
setup and fixtures was necessary, but overall, it feels like a large win.