Refactor, comment and generally clean up code
Got a good feeling about this. Despite having started off as being a giant
500+ line script with a bunch of hacks to get a single test suite converted,
it was fairly straightforward to rewrite it and give a high-level overview of
how this tool works.

Given a package name X
  * find all of its tests, and its child tests
  * create a ginkgo test suite file with `ginkgo bootstrap`
  * for each of its tests, parse its AST and find the Test funcs
  * add import statements for ginkgo, removing the "testing" import
  * add a top level Describe and rewrite each test as an It()
  * find the remaining usages of *testing.T in the file and replace with GinkgoT
  * save the file back out

Granted, there are quite a few edge cases, but that's what tests are for.